Refactored Restserver.go
